+++ lib/src/option.dart
@@ -38,24 +38,33 @@ class Option<Some> extends Equatable {
   }
 
   /// Returns `true` if the option is `Some`.
-  bool isSome() {
-    return _some != null;
-  }
+  bool get isSome => _some != null;
 
   /// Returns `true` if the option is `None`.
-  bool isNone() {
-    return _some == null;
-  }
+  bool get isNone => _some == null;
 
   /// Invokes either the `someop` or the `noneop` depending on the option.
   ///
   /// This is an attempt at providing something similar to the Rust `match`
   /// expression, which makes it easy to handle both cases at once.
+  ///
+  /// See also [when] for another way to achieve the same behavior.
   void match(void Function(Some) someop, void Function() noneop) {
     if (_some != null) {
       someop(_some);
     } else {
       noneop();
+    }
+  }
+
+  /// Invokes either `some` or `none` depending on the option.
+  ///
+  /// Identical to [match] except that the arguments are named.
+  void when({void Function(Some) some, void Function() none}) {
+    if (_some != null) {
+      some(_some);
+    } else {
+      none();
     }
   }
 
@@ -195,10 +204,10 @@ class Option<Some> extends Equatable {
   /// Returns `Some` if exactly one of `this`, `optb` is `Some`, otherwise
   /// returns `None`.
   Option<Some> xor(Option<Some> optb) {
-    if (_some != null && optb.isNone()) {
+    if (_some != null && optb.isNone) {
       return this;
     }
-    if (_some == null && optb.isSome()) {
+    if (_some == null && optb.isSome) {
       return optb;
     }
     return Option.none();