+++ CHANGELOG.md
@@ -5,6 +5,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file follows the convention described at
 [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+### Changed
+- Changed the `is` methods to be getters, breaking the old API.
+- Extend `Equatable` in both `Option` and `Result`.
+- Override the equals operator (`==`) in both `Option` and `Result`.
+
 ## [1.0.0] - 2020-03-09
 ### Added
 - Initial version